requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3

# Deep Learning
torch==2.0.1
torchvision==0.15.2

# Image Processing
Pillow==10.0.0
opencv-python==4.8.0.76
scikit-image==0.21.0
imageio==2.31.2
numpy==1.24.3

# Scientific Computing
scipy==1.11.2
matplotlib==3.7.2

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0

# Optional - For API enhancements
flask-cors==4.0.0
flask-limiter==3.3.1
